requirements.txt
tkinter
mysql-connector-python==8.0.33
bcrypt==4.0.1
python-docx==0.8.11
Pillow==10.0.0
secrets 